moved DrtServiceQualityProbeParams to different module
