+++ contribs/drt/src/test/java/org/matsim/contrib/drt/optimizer/servicequality/DrtServiceQualityProbeParamsTest.java
@@ -1,4 +1,4 @@
-package org.matsim.contrib.drt.optimizer.insertion.parallel;
+package org.matsim.contrib.drt.optimizer.servicequality;
 
 import org.junit.jupiter.api.Test;
 import org.matsim.contrib.drt.optimizer.insertion.repeatedselective.RepeatedSelectiveInsertionSearchParams;
